feat: Implement comprehensive late fee calculation, policy management, and discount definitions within the fee payment system.

+++ backend/src/main/java/com/school/backend/fee/dto/FeeStatsDto.java
@@ -4,13 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class FeeStatsDto {
-    private long todayCollection;
-    private long pendingDues;
+    private BigDecimal todayCollection;
+    private BigDecimal pendingDues;
     private long totalStudents;
 }
